requirements.txt
# Plotting packages
matplotlib==3.9.0
seaborn==0.13.2

# Data wrangling packages
numpy==1.26.4
pandas==2.2.2

# Open AI packages 
openai==1.33.0
python-dotenv==1.0.1

# Front-end packages
streamlit==1.35.0